cleanup old files and rename stuff
