+++ core/CMakeLists.txt
@@ -22,14 +22,10 @@ detray_add_library( detray_core core
    "include/detray/core/detail/container_view.hpp"
    "include/detray/core/detail/data_context.hpp"
    "include/detray/core/detail/multi_store.hpp"
-   "include/detray/core/detail/new_tuple_container.hpp"
    "include/detray/core/detail/single_store.hpp"
-   "include/detray/core/detail/tuple_array_container.hpp"
    "include/detray/core/detail/tuple_container.hpp"
-   "include/detray/core/detail/type_registry.hpp"
    "include/detray/core/detector.hpp"
    "include/detray/core/detector_kernel.hpp"
-   "include/detray/core/surfaces_finder.hpp"
    # definitions include(s)
    "include/detray/definitions/containers.hpp"
    "include/detray/definitions/cuda_definitions.hpp"
@@ -93,9 +89,6 @@ detray_add_library( detray_core core
    "include/detray/propagator/rk_stepper.hpp"
    # surface finder include(s)
    "include/detray/surface_finders/brute_force_finder.hpp"
-   "include/detray/surface_finders/grid2_finder.hpp"
-   "include/detray/surface_finders/neighborhood_kernel.hpp"
-   # surface finder include(s)
    "include/detray/surface_finders/grid/detail/axis_binning.hpp"
    "include/detray/surface_finders/grid/detail/axis_helpers.hpp"
    "include/detray/surface_finders/grid/detail/axis_shape.hpp"
@@ -108,6 +101,7 @@ detray_add_library( detray_core core
    "include/detray/surface_finders/grid/grid.hpp"
    "include/detray/surface_finders/grid/populator.hpp"
    "include/detray/surface_finders/grid/serializer.hpp"
+   "include/detray/surface_finders/neighborhood_kernel.hpp"
    # tools include(s)
    "include/detray/tools/associator.hpp"
    "include/detray/tools/bin_association.hpp"
@@ -136,5 +130,6 @@ detray_add_library( detray_core core
    "include/detray/utils/quadratic_equation.hpp"
    "include/detray/utils/ratio.hpp"
    "include/detray/utils/statistics.hpp"
+   "include/detray/utils/type_registry.hpp"
    "include/detray/utils/type_traits.hpp" )
 target_link_libraries( detray_core INTERFACE detray_io vecmem::core detray::Thrust)
