--- tests/unit_tests/array/array_surfaces_finder.cpp
@@ -1,9 +1,0 @@
-/* Detray library, part of the ACTS project (R&D line)
- *
- * (c) 2021 CERN for the benefit of the ACTS project
- *
- * Mozilla Public License Version 2.0
- */
-
-#include "detray/plugins/algebra/array_definitions.hpp"
-#include "tests/common/surfaces_finder.inl"
