+++ tests/unit_tests/smatrix/CMakeLists.txt
@@ -47,8 +47,7 @@ detray_add_test( smatrix
    "smatrix_rectangle2D.cpp"
    "smatrix_ring2D.cpp"
    "smatrix_single3D.cpp"
-   "smatrix_stepper.cpp" 
-   "smatrix_surfaces_finder.cpp"
+   "smatrix_stepper.cpp"
    "smatrix_telescope_detector.cpp"
    "smatrix_toy_geometry.cpp"
    "smatrix_track.cpp"
